+++ src/com/dndsuite/core/json/parsers/subevents/Calculation.java
@@ -2,10 +2,10 @@ package com.dndsuite.core.json.parsers.subevents;
 
 public interface Calculation {
 
-	public abstract void addBonus(int bonus);
+	public abstract void addBonus(long bonus);
 
-	public abstract void setTo(int set);
+	public abstract void setTo(long set);
 
-	public abstract int get();
+	public abstract long get();
 
 }
